多主多从，不分表配置

+++ src/main/java/com/skh/service/UserService.java
@@ -6,5 +6,5 @@ public interface UserService {
 
     User getUser(int id);
 
-    Integer saveUser(User user);
+    Long saveUser(User user);
 }
